Change permissions

+++ convert_to_git.sh
@@ -1,22 +1,30 @@
 #!/bin/bash
 
-# use HTTPS by default to allow non-members to clone the repos
-if [ $# -gt 0 ]; then 
-	GITHUB_URL=git@github.com:OHBA-analysis
-else 
-	GITHUB_URL=https://github.com/OHBA-analysis
-fi
+DEBUG=0
+GITHUB_URL=https://github.com/OHBA-analysis
 
 replace_repo () {
 	if [ -d "$1" ]; then
-		rm -rf "$1"
-		git clone "${GITHUB_URL}/$1"
+		if (( $DEBUG )); then 
+			echo "Cloning repository '${GITHUB_URL}/$1' into folder '$1'"
+		else 
+			rm -rf "$1"
+			git clone "${GITHUB_URL}/$1"
+		fi
 	fi
 } 
 
-read -r -p "This will delete all OSL folders and clone from git. Use with caution! Continue? [y/N] " response
+read -r -p 'This will delete all OSL folders and clone from git. Use with caution! Continue? [y/N] ' response
 case "$response" in
     [yY][eE][sS]|[yY]) 
+
+		read -r -p 'Is this computer associated with a GitHub account, which has access to the OHBA Analysis Group repositories? [y/N] ' response
+		case "$response" in
+    		[yY][eE][sS]|[yY])
+				GITHUB_URL=git@github.com:OHBA-analysis
+				;;
+		esac
+
 		cd ..
 		replace_repo osl-core
 		replace_repo ohba-external
@@ -26,6 +34,6 @@ case "$response" in
 		cd osl-core
         ;;
     *)
-        echo "Cancelled"
+        echo 'Cancelled'
         ;;
 esac
